% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATC_classification.R
\docType{data}
\name{ATC_CATEGORIES}
\alias{ATC_CATEGORIES}
\title{ATC Categories}
\format{
A list with 14 elements, each representing an ATC category
}
\usage{
ATC_CATEGORIES
}
\description{
A list containing all Anatomical Therapeutic Chemical (ATC) classification categories
with their codes and full names.
}
\keyword{datasets}
