% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATC_classification.R
\name{build_atc_query}
\alias{build_atc_query}
\title{Build ATC Query}
\usage{
build_atc_query(schema, atc_code)
}
\arguments{
\item{schema}{Character string with the database schema name}

\item{atc_code}{Character string with the ATC code}
}
\value{
A character string containing the SQL query
}
\description{
Builds a SQL query for retrieving prescription statistics for a specific ATC code.
}
\keyword{internal}
