% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_categories.R
\docType{data}
\name{DISEASE_CATEGORIES}
\alias{DISEASE_CATEGORIES}
\title{Disease Categories Based on ICD-10 Codes}
\format{
A list of lists where each sublist contains:
\describe{
\item{name}{Full name of the disease category}
\item{code_range}{Vector of two elements indicating start and end ICD-10 codes}
}
}
\usage{
DISEASE_CATEGORIES
}
\description{
A list containing standardized disease categories with their ICD-10 code ranges
and descriptive names.
}
\keyword{datasets}
