% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_functions.R
\name{get_total_patients}
\alias{get_total_patients}
\title{Get Total Patient Count}
\usage{
get_total_patients(conn, cdm_schema)
}
\arguments{
\item{conn}{A DBI connection object to the database}

\item{cdm_schema}{Character string specifying the CDM schema name}
}
\value{
Integer representing the total count of patients
}
\description{
Calculates the total number of unique patients in the database
}
\examples{
\dontrun{
conn <- DBI::dbConnect(...)
total <- get_total_patients(conn, "cdm_schema")
}
}
