% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_atc_code}
\alias{validate_atc_code}
\title{Validate ATC Code}
\usage{
validate_atc_code(atc_code)
}
\arguments{
\item{atc_code}{Character string of ATC code}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Internal function to validate ATC code
}
\keyword{internal}
