% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_category}
\alias{validate_category}
\title{Validate Category Name}
\usage{
validate_category(category)
}
\arguments{
\item{category}{Character string of category name}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Internal function to validate disease category name
}
\keyword{internal}
