% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_connection}
\alias{validate_connection}
\title{Validate Database Connection}
\usage{
validate_connection(conn)
}
\arguments{
\item{conn}{A DBI connection object}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Internal function to validate database connection object
}
\keyword{internal}
