% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_connection_db_connector}
\alias{validate_connection_db_connector}
\title{Validate DatabaseConnector Connection}
\usage{
validate_connection_db_connector(conn)
}
\arguments{
\item{conn}{A DatabaseConnector connection object}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Internal function to validate DatabaseConnector connection object
}
\keyword{internal}
