% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_schema}
\alias{validate_schema}
\title{Validate Schema Name}
\usage{
validate_schema(schema)
}
\arguments{
\item{schema}{Character string of schema name}
}
\value{
TRUE if valid, throws error if invalid
}
\description{
Internal function to validate schema name
}
\keyword{internal}
